+++ scripts/export-dos-participants.py
@@ -8,7 +8,7 @@ import sys
 sys.path.insert(0, '.')
 
 from docopt import docopt
-from dmscripts.helpers.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env_helpers import get_api_endpoint_from_stage
 from dmscripts.export_dos_suppliers import (
     FRAMEWORK_SLUG, find_services_by_lot, make_fields_from_content_questions, write_csv
 )
