+++ scripts/generate-agreement-signature-pages.py
@@ -27,7 +27,7 @@ import tempfile
 sys.path.insert(0, '.')
 
 from docopt import docopt
-from dmscripts.helpers.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env_helpers import get_api_endpoint_from_stage
 from dmscripts.export_framework_applicant_details import find_suppliers_with_details
 from dmscripts.generate_agreement_signature_pages import render_html_for_successful_suppliers, \
     render_pdf_for_each_html_page
