+++ scripts/mark-definite-framework-results.py
@@ -38,7 +38,7 @@ import getpass
 import json
 
 
-from dmscripts.helpers.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env_helpers import get_api_endpoint_from_stage
 from dmapiclient import DataAPIClient
 from dmscripts.mark_definite_framework_results import mark_definite_framework_results
 from dmscripts.logging import configure_logger, INFO as loglevel_INFO, DEBUG as loglevel_DEBUG
