+++ scripts/set-search-alias.py
@@ -10,7 +10,7 @@ import sys
 from docopt import docopt
 
 sys.path.insert(0, '.')
-from dmscripts.helpers.env import get_api_endpoint_from_stage
+from dmscripts.helpers.env_helpers import get_api_endpoint_from_stage
 from dmapiclient import SearchAPIClient
